Add index.html
This file will be the beginning of my first project with the odin
project. What is being created is a what will be a recipe book.
From here on out I will be creating not only the website but also
pushing commits through git within the command line. After I am done
with the project I will express what differences I have noticed with
my progression.