Finish Iteration 3: Recipe Page Content
Created some new additions to sangria.html: Added an image of the recipe
being created, a heading description, a paragraph describing what's being
made, an unordered list for ingredients was made including all items
needed, and an ordered list containing the steps needed to finish the
recipe.

This iteration contains the bulk of the text needed to finish this page.

Now we can continue to Iteration 4 of the project.